CBG-3696: Allow user_xattr_key to be removed with an empty-string POST/upsert (#6627)
* Allow user_xattr_key to be removed with an empty-string POST/upsert

* Move user xattr tests into their own package and organise by functionality

* Add new TestUserXattrSetUnsetDBConfig test for empty string POST/Upsert and nil on PUT

* Run addlicense

* Force new test to run with persistent config

* Address PR comments
