+++ rest/config_database.go
@@ -173,7 +173,7 @@ func DefaultDbConfig(sc *StartupConfig) *DbConfig {
 		Replications:                     nil,
 		ServeInsecureAttachmentTypes:     base.BoolPtr(false),
 		QueryPaginationLimit:             base.IntPtr(db.DefaultQueryPaginationLimit),
-		UserXattrKey:                     "",
+		UserXattrKey:                     nil,
 		ClientPartitionWindowSecs:        base.IntPtr(int(base.DefaultClientPartitionWindow.Seconds())),
 		JavascriptTimeoutSecs:            base.Uint32Ptr(base.DefaultJavascriptTimeoutSecs),
 		Suspendable:                      base.BoolPtr(sc.IsServerless()),
