+++ rest/user_api_test.go
@@ -550,58 +550,6 @@ func TestUserAndRoleResponseContentType(t *testing.T) {
 	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
 }
 
-func TestUserXattrsRawGet(t *testing.T) {
-	if !base.TestUseXattrs() {
-		t.Skip("Test requires xattrs to be enabled")
-	}
-
-	if !base.IsEnterpriseEdition() {
-		t.Skipf("test is EE only - user xattrs")
-	}
-
-	docKey := t.Name()
-	xattrKey := "xattrKey"
-
-	rt := NewRestTester(t, &RestTesterConfig{
-		DatabaseConfig: &DatabaseConfig{
-			DbConfig: DbConfig{
-				AutoImport:   true,
-				UserXattrKey: xattrKey,
-			},
-		},
-	})
-	defer rt.Close()
-
-	userXattrStore, ok := base.AsUserXattrStore(rt.GetSingleDataStore())
-	if !ok {
-		t.Skip("Test requires Couchbase Bucket")
-	}
-
-	resp := rt.SendAdminRequest("PUT", "/{{.keyspace}}/"+docKey, "{}")
-	RequireStatus(t, resp, http.StatusCreated)
-	require.NoError(t, rt.WaitForPendingChanges())
-
-	_, err := userXattrStore.WriteUserXattr(docKey, xattrKey, "val")
-	assert.NoError(t, err)
-
-	err = rt.WaitForCondition(func() bool {
-		return rt.GetDatabase().DbStats.SharedBucketImportStats.ImportCount.Value() == 1
-	})
-
-	resp = rt.SendAdminRequest("GET", "/{{.keyspace}}/_raw/"+docKey, "")
-	RequireStatus(t, resp, http.StatusOK)
-
-	var RawReturn struct {
-		Meta struct {
-			Xattrs map[string]interface{} `json:"xattrs"`
-		} `json:"_meta"`
-	}
-
-	err = json.Unmarshal(resp.BodyBytes(), &RawReturn)
-
-	assert.Equal(t, "val", RawReturn.Meta.Xattrs[xattrKey])
-}
-
 func TestObtainUserChannelsForDeletedRoleCasFail(t *testing.T) {
 	if !base.UnitTestUrlIsWalrus() {
 		t.Skip("Skip LeakyBucket test when running in integration")
@@ -1182,7 +1130,7 @@ func TestRemovingUserXattr(t *testing.T) {
 			rt := NewRestTester(t, &RestTesterConfig{
 				DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
 					AutoImport:   testCase.autoImport,
-					UserXattrKey: xattrKey,
+					UserXattrKey: &xattrKey,
 				}},
 				SyncFn: `
 			function (doc, oldDoc, meta){
@@ -1243,292 +1191,6 @@ func TestRemovingUserXattr(t *testing.T) {
 			assert.Equal(t, uint64(3), syncData2.Channels[channelName].Seq)
 		})
 	}
-}
-
-func TestUserXattrRevCache(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAll)
-	if base.UnitTestUrlIsWalrus() {
-		t.Skip("This test only works against Couchbase Server")
-	}
-
-	if !base.TestUseXattrs() {
-		t.Skip("This test only works with XATTRS enabled")
-	}
-
-	if !base.IsEnterpriseEdition() {
-		t.Skipf("test is EE only - user xattrs")
-	}
-
-	docKey := t.Name()
-	xattrKey := "channels"
-	channelName := []string{"ABC", "DEF"}
-	tb := base.GetTestBucket(t)
-	defer tb.Close(base.TestCtx(t))
-	syncFn := `function (doc, oldDoc, meta){
-				if (meta.xattrs.channels !== undefined){
-					channel(meta.xattrs.channels);
-					console.log(JSON.stringify(meta));
-				}
-			}`
-
-	// Sync function to set channel access to a channels UserXattrKey
-	rt := NewRestTester(t, &RestTesterConfig{
-		CustomTestBucket: tb.NoCloseClone(),
-		DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
-			AutoImport:   true,
-			UserXattrKey: xattrKey,
-		}},
-		SyncFn: syncFn,
-	})
-	defer rt.Close()
-
-	rt2 := NewRestTester(t, &RestTesterConfig{
-		CustomTestBucket: tb.NoCloseClone(),
-
-		DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
-			AutoImport:   true,
-			UserXattrKey: xattrKey,
-		}},
-		SyncFn: syncFn,
-	})
-	defer rt2.Close()
-
-	dataStore := rt2.GetSingleDataStore()
-	userXattrStore, ok := base.AsUserXattrStore(dataStore)
-	if !ok {
-		t.Skip("Test requires Couchbase Bucket")
-	}
-
-	ctx := rt2.Context()
-	a := rt2.ServerContext().Database(ctx, "db").Authenticator(ctx)
-	userABC, err := a.NewUser("userABC", "letmein", channels.BaseSetOf(t, "ABC"))
-	require.NoError(t, err)
-	require.NoError(t, a.Save(userABC))
-
-	userDEF, err := a.NewUser("userDEF", "letmein", channels.BaseSetOf(t, "DEF"))
-	require.NoError(t, err)
-	require.NoError(t, a.Save(userDEF))
-
-	resp := rt.SendAdminRequest("PUT", "/{{.keyspace}}/"+docKey, `{}`)
-	RequireStatus(t, resp, http.StatusCreated)
-	require.NoError(t, rt.WaitForPendingChanges())
-	_, err = userXattrStore.WriteUserXattr(docKey, xattrKey, "DEF")
-	assert.NoError(t, err)
-
-	_, err = rt.WaitForChanges(1, "/{{.keyspace}}/_changes", "userDEF", false)
-	assert.NoError(t, err)
-
-	resp = rt2.SendUserRequest("GET", "/{{.keyspace}}/"+docKey, ``, "userDEF")
-	RequireStatus(t, resp, http.StatusOK)
-
-	// Add channel ABC to the userXattr
-	_, err = userXattrStore.WriteUserXattr(docKey, xattrKey, channelName)
-	assert.NoError(t, err)
-
-	// wait for import of the xattr change on both nodes
-	_, err = rt.WaitForChanges(1, "/{{.keyspace}}/_changes", "userABC", false)
-	assert.NoError(t, err)
-	_, err = rt2.WaitForChanges(1, "/{{.keyspace}}/_changes", "userABC", false)
-	assert.NoError(t, err)
-
-	// GET the doc with userABC to ensure it is accessible on both nodes
-	resp = rt2.SendUserRequest("GET", "/{{.keyspace}}/"+docKey, ``, "userABC")
-	assert.Equal(t, resp.Code, http.StatusOK)
-	resp = rt.SendUserRequest("GET", "/{{.keyspace}}/"+docKey, ``, "userABC")
-	assert.Equal(t, resp.Code, http.StatusOK)
-}
-
-func TestUserXattrDeleteWithRevCache(t *testing.T) {
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAll)
-	if base.UnitTestUrlIsWalrus() {
-		t.Skip("This test only works against Couchbase Server")
-	}
-
-	if !base.TestUseXattrs() {
-		t.Skip("This test only works with XATTRS enabled")
-	}
-
-	if !base.IsEnterpriseEdition() {
-		t.Skipf("test is EE only - user xattrs")
-	}
-
-	// Sync function to set channel access to a channels UserXattrKey
-	syncFn := `
-			function (doc, oldDoc, meta){
-				if (meta.xattrs.channels !== undefined){
-					channel(meta.xattrs.channels);
-					console.log(JSON.stringify(meta));
-				}
-			}`
-
-	docKey := t.Name()
-	xattrKey := "channels"
-	tb := base.GetTestBucket(t)
-	defer tb.Close(base.TestCtx(t))
-
-	rt := NewRestTester(t, &RestTesterConfig{
-		CustomTestBucket: tb.NoCloseClone(),
-		DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
-			AutoImport:   true,
-			UserXattrKey: xattrKey,
-		}},
-		SyncFn: syncFn,
-	})
-	defer rt.Close()
-
-	rt2 := NewRestTester(t, &RestTesterConfig{
-		CustomTestBucket: tb.NoCloseClone(),
-		DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
-			AutoImport:   true,
-			UserXattrKey: xattrKey,
-		}},
-		SyncFn: syncFn,
-	})
-	defer rt2.Close()
-
-	dataStore := rt2.GetSingleDataStore()
-	userXattrStore, ok := base.AsUserXattrStore(dataStore)
-	if !ok {
-		t.Skip("Test requires Couchbase Bucket")
-	}
-
-	ctx := rt2.Context()
-	a := rt2.ServerContext().Database(ctx, "db").Authenticator(ctx)
-
-	userDEF, err := a.NewUser("userDEF", "letmein", channels.BaseSetOf(t, "DEF"))
-	require.NoError(t, err)
-	require.NoError(t, a.Save(userDEF))
-
-	resp := rt.SendAdminRequest("PUT", "/{{.keyspace}}/"+docKey, `{}`)
-	RequireStatus(t, resp, http.StatusCreated)
-	require.NoError(t, rt.WaitForPendingChanges())
-
-	// Write DEF to the userXattrStore to give userDEF access
-	_, err = userXattrStore.WriteUserXattr(docKey, xattrKey, "DEF")
-	assert.NoError(t, err)
-
-	_, err = rt.WaitForChanges(1, "/{{.keyspace}}/_changes", "userDEF", false)
-	assert.NoError(t, err)
-
-	resp = rt2.SendUserRequest("GET", "/{{.keyspace}}/"+docKey, ``, "userDEF")
-	RequireStatus(t, resp, http.StatusOK)
-
-	// Delete DEF from the userXattr, removing the doc from channel DEF
-	_, err = userXattrStore.DeleteUserXattr(docKey, xattrKey)
-	assert.NoError(t, err)
-
-	// wait for import of the xattr change on both nodes
-	_, err = rt.WaitForChanges(1, "/{{.keyspace}}/_changes", "userDEF", false)
-	assert.NoError(t, err)
-	_, err = rt2.WaitForChanges(1, "/{{.keyspace}}/_changes", "userDEF", false)
-	assert.NoError(t, err)
-
-	// GET the doc with userDEF on both nodes to ensure userDEF no longer has access
-	resp = rt2.SendUserRequest("GET", "/{{.keyspace}}/"+docKey, ``, "userDEF")
-	assert.Equal(t, resp.Code, http.StatusForbidden)
-	resp = rt.SendUserRequest("GET", "/{{.keyspace}}/"+docKey, ``, "userDEF")
-	assert.Equal(t, resp.Code, http.StatusForbidden)
-}
-
-func TestUserXattrAvoidRevisionIDGeneration(t *testing.T) {
-	if base.UnitTestUrlIsWalrus() {
-		t.Skip("This test only works against Couchbase Server")
-	}
-
-	if !base.TestUseXattrs() {
-		t.Skip("This test only works with XATTRS enabled")
-	}
-
-	if !base.IsEnterpriseEdition() {
-		t.Skipf("test is EE only - user xattrs")
-	}
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAll)
-
-	docKey := t.Name()
-	xattrKey := "myXattr"
-	channelName := "testChan"
-
-	// Sync function to set channel access to whatever xattr is
-	rt := NewRestTester(t, &RestTesterConfig{
-		DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
-			AutoImport:   true,
-			UserXattrKey: xattrKey,
-		}},
-		SyncFn: `
-			function (doc, oldDoc, meta){
-				if (meta.xattrs.myXattr !== undefined){
-					channel(meta.xattrs.myXattr);
-					console.log(JSON.stringify(meta));
-				}
-			}`,
-	})
-
-	defer rt.Close()
-
-	dataStore := rt.GetSingleDataStore()
-	userXattrStore, ok := base.AsUserXattrStore(dataStore)
-	if !ok {
-		t.Skip("Test requires Couchbase Bucket")
-	}
-
-	subdocXattrStore, ok := base.AsSubdocXattrStore(dataStore)
-	require.True(t, ok)
-
-	// Initial PUT
-	resp := rt.SendAdminRequest("PUT", "/{{.keyspace}}/"+docKey, `{}`)
-	RequireStatus(t, resp, http.StatusCreated)
-
-	require.NoError(t, rt.WaitForPendingChanges())
-
-	ctx := base.TestCtx(t)
-	// Get current sync data
-	var syncData db.SyncData
-	_, err := subdocXattrStore.SubdocGetXattr(ctx, docKey, base.SyncXattrName, &syncData)
-	assert.NoError(t, err)
-
-	docRev, err := rt.GetSingleTestDatabaseCollection().GetRevisionCacheForTest().Get(base.TestCtx(t), docKey, syncData.CurrentRev, true, false)
-	assert.NoError(t, err)
-	assert.Equal(t, 0, len(docRev.Channels.ToArray()))
-	assert.Equal(t, syncData.CurrentRev, docRev.RevID)
-
-	// Write xattr to trigger import of user xattr
-	_, err = userXattrStore.WriteUserXattr(docKey, xattrKey, channelName)
-	assert.NoError(t, err)
-
-	// Wait for import
-	err = rt.WaitForCondition(func() bool {
-		return rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value() == 1
-	})
-	assert.NoError(t, err)
-
-	// Ensure import worked and sequence incremented but that sequence did not
-	var syncData2 db.SyncData
-	_, err = subdocXattrStore.SubdocGetXattr(ctx, docKey, base.SyncXattrName, &syncData2)
-	assert.NoError(t, err)
-
-	docRev2, err := rt.GetSingleTestDatabaseCollection().GetRevisionCacheForTest().Get(base.TestCtx(t), docKey, syncData.CurrentRev, true, false)
-	assert.NoError(t, err)
-	assert.Equal(t, syncData2.CurrentRev, docRev2.RevID)
-
-	assert.Equal(t, syncData.CurrentRev, syncData2.CurrentRev)
-	assert.True(t, syncData2.Sequence > syncData.Sequence)
-	assert.Equal(t, []string{channelName}, syncData2.Channels.KeySet())
-	assert.Equal(t, syncData2.Channels.KeySet(), docRev2.Channels.ToArray())
-
-	err = rt.GetSingleDataStore().Set(docKey, 0, nil, []byte(`{"update": "update"}`))
-	assert.NoError(t, err)
-
-	err = rt.WaitForCondition(func() bool {
-		log.Printf("Import count is: %v", rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value())
-		return rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value() == 2
-	})
-	assert.NoError(t, err)
-
-	var syncData3 db.SyncData
-	_, err = subdocXattrStore.SubdocGetXattr(ctx, docKey, base.SyncXattrName, &syncData2)
-	assert.NoError(t, err)
-
-	assert.NotEqual(t, syncData2.CurrentRev, syncData3.CurrentRev)
 }
 
 func TestGetUserCollectionAccess(t *testing.T) {
